[DD-546] Permanent Push: Add Sign-up CTA to docs/2.0/first-steps (#6691)
* remove experiment logic form code base

* bring functionality of btn toggle to main

* remove experiment styles and update class names
