+++ src/js/experiments/index.js
@@ -1,6 +1,5 @@
 import forceAll from './forceAll';
 import languageGuides from './languageGuides';
-import signupLinksCta from './signupLinksCta';
 import './snippetFeedback';
 import troubleShootingCommonIssues from './troubleshootingCommonIssues';
 import popupCta from './popupCta';
@@ -8,7 +7,6 @@ import popupCta from './popupCta';
 export default {
   forceAll,
   languageGuides,
-  signupLinksCta,
   troubleShootingCommonIssues,
   popupCta,
 };
